refactor: 파일 분리 및 products/size로 변경
