+++ app/src/main/java/woowacourse/shopping/productdetail/DetailProductUiState.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.productdetail
+
+import woowacourse.shopping.productlist.ProductUiModel
+import woowacourse.shopping.productlist.ViewedProductUiModel
+
+data class DetailProductUiState(
+    val productUiModel: ProductUiModel,
+    val lastViewedProductUiModel: ViewedProductUiModel?,
+)
