+++ app/src/main/java/woowacourse/shopping/productlist/ProductListActivity.kt
@@ -20,6 +20,6 @@ class ProductListActivity : ComponentActivity() {
     }
 
     companion object {
-        const val EXTRA_PRODUCT_ID = "productId"
+        const val PRODUCT_ID = "productId"
     }
 }
