+++ app/src/main/java/woowacourse/shopping/productlist/ProductListUiState.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.productlist
+
+data class ProductListUiState(
+    val productUiModels: List<ProductUiModel>,
+    val viewedProductUiModels: List<ViewedProductUiModel>,
+    val cartItemCount: Int,
+    val isLoadMoreEnabled: Boolean,
+)
